refactor: 정답을 판단하는 wordResolver 패키지 이동
