+++ src/main/kotlin/wordle/Wordle.kt
@@ -2,11 +2,12 @@ package wordle
 
 import wordle.domain.Results
 import wordle.domain.Word
+import wordle.domain.WordResolver
 
 class Wordle(
     private val resolver: WordResolver
 ) {
     fun round(word: Word): Results {
-        return Results(resolver.check(word))
+        return resolver.check(word)
     }
 }
